+++ lizard_graph/admin.py
@@ -10,6 +10,9 @@ from lizard_graph.models import HorizontalBarGraphItem
 
 class GraphItemInline(admin.TabularInline):
     model = GraphItem
+    fields = ('location', 'parameter', 'module', 'index', 'graph_type',
+              'value', 'color', 'color_outside', 'line_width',
+              'line_style', 'label', )
 
 
 class PredefinedGraphAdmin(admin.ModelAdmin):